Safari-611.1.10/LayoutTests/fast/forms/ValidityState-tooLong-input-expected.txt

Tests for tooLong flag with <input> elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


No maxlength and no value
PASS input.validity.tooLong is false

Non-dirty value
PASS input.value.length is 5
PASS input.validity.tooLong is false
PASS input.value.length is 6
PASS input.validity.tooLong is false

Dirty value and longer than maxLength
PASS input.value.length is 4
PASS input.validity.tooLong is true
PASS input.validity.tooLong is false

Sets a value via DOM property
PASS input.validity.tooLong is false

Disabled
PASS input.validity.tooLong is false

Grapheme length is not greater than maxLength though character length is greater
PASS input.validity.tooLong is false

Change the type with a too long value
PASS input.validity.tooLong is true
PASS input.type = "number"; input.validity.tooLong is false
PASS successfullyParsed is true

TEST COMPLETE
